# information-srv/client/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Contain Generic Information.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Information API
  version: "1.0"
servers:
- url: /
paths:
  /internal/address:
    get:
      description: Retrieve a complete list of all available addresses.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Addresses
      tags:
      - Address
  /internal/address/province:
    get:
      description: Retrieve a complete list of all provinces.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_province_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Provinces
      tags:
      - Address
  /internal/address/province/{province}:
    get:
      description: Retrieve a list of addresses associated with a specified Province.
      parameters:
      - description: The Province to retrieve addresses
        explode: false
        in: path
        name: province
        required: true
        schema:
          type: string
        style: simple
      - description: Language preference for response data. Omit for Thai (default)
          or specify 'en' for English
        explode: true
        in: query
        name: lang
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Addresses by Province
      tags:
      - Address
  /internal/address/zip-code/{zipCode}:
    get:
      description: Retrieve a list of addresses associated with a specified Zip Code.
      parameters:
      - description: The Zip Code to retrieve addresses
        explode: false
        in: path
        name: zipCode
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Addresses by Zip Code
      tags:
      - Address
  /internal/bank:
    get:
      description: Retrieve a complete list of all banks.
      parameters:
      - description: Id
        explode: true
        in: query
        name: id
        required: false
        schema:
          type: string
        style: form
      - description: ShortName
        explode: true
        in: query
        name: shortName
        required: false
        schema:
          type: string
        style: form
      - description: Code
        explode: true
        in: query
        name: code
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_bank_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Banks
      tags:
      - Bank
  /internal/bank-branch:
    get:
      description: Retrieve a complete list of all bank branches.
      parameters:
      - description: BankCode
        explode: true
        in: query
        name: bankCode
        required: false
        schema:
          type: string
        style: form
      - description: BankBranchCode
        explode: true
        in: query
        name: bankBranchCode
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_bank_branch_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All BankBranches
      tags:
      - BankBranch
...
  /secure/address:
    get:
      description: Retrieve a complete list of all available addresses.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Addresses
      tags:
      - Address
  /secure/address/province:
    get:
      description: Retrieve a complete list of all provinces.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_province_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Provinces
      tags:
      - Address
  /secure/address/province/{province}:
    get:
      description: Retrieve a list of addresses associated with a specified Province.
      parameters:
      - description: The Province to retrieve addresses
        explode: false
        in: path
        name: province
        required: true
        schema:
          type: string
        style: simple
      - description: Language preference for response data. Omit for Thai (default)
          or specify 'en' for English
        explode: true
        in: query
        name: lang
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Addresses by Province
      tags:
      - Address
  /secure/address/zip-code/{zipCode}:
    get:
      description: Retrieve a list of addresses associated with a specified Zip Code.
      parameters:
      - description: The Zip Code to retrieve addresses
        explode: false
        in: path
        name: zipCode
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_address_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get Addresses by Zip Code
      tags:
      - Address
  /secure/bank:
    get:
      description: Retrieve a complete list of all banks.
      parameters:
      - description: Id
        explode: true
        in: query
        name: id
        required: false
        schema:
          type: string
        style: form
      - description: ShortName
        explode: true
        in: query
        name: shortName
        required: false
        schema:
          type: string
        style: form
      - description: Code
        explode: true
        in: query
        name: code
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_bank_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Banks
      tags:
      - Bank
  /secure/bank-branch:
    get:
      description: Retrieve a complete list of all bank branches.
      parameters:
      - description: BankCode
        explode: true
        in: query
        name: bankCode
        required: false
        schema:
          type: string
        style: form
      - description: BankBranchCode
        explode: true
        in: query
        name: bankBranchCode
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_bank_branch_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All BankBranches
      tags:
      - BankBranch
...
components:
  schemas:
    address.Address:
      example:
        sub_district_en: sub_district_en
        district_th: district_th
        province_en: province_en
        province_th: province_th
        sub_district_th: sub_district_th
        district_en: district_en
        zip_code: zip_code
      properties:
        district_en:
          type: string
        district_th:
          type: string
        province_en:
          type: string
        province_th:
          type: string
        sub_district_en:
          type: string
        sub_district_th:
          type: string
        zip_code:
          type: string
      type: object
    address.Province:
      example:
        code: code
        province_en: province_en
        iso_code: iso_code
        province_th: province_th
      properties:
        code:
          type: string
        iso_code:
          type: string
        province_en:
          type: string
        province_th:
          type: string
      type: object
    bank.Bank:
      example:
        icon_url: icon_url
        code: code
        name: name
        name_th: name_th
        short_name: short_name
        id: id
      properties:
        code:
          type: string
        icon_url:
          type: string
        id:
          type: string
        name:
          type: string
        name_th:
          type: string
        short_name:
          type: string
      type: object
    bankBranch.BankBranch:
      example:
        bank_code: bank_code
        bank_branch_code: bank_branch_code
        branch_name: branch_name
        id: id
      properties:
        bank_branch_code:
          type: string
        bank_code:
          type: string
        branch_name:
          type: string
        id:
          type: string
      type: object
...
    pkg.Response:
      example:
        detail: detail
        title: title
        status: 6
      properties:
        detail:
          type: string
        status:
          type: integer
        title:
          type: string
      type: object
...
    _internal_address_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/address.Address'
            type: array
        type: object
      example:
        data:
        - sub_district_en: sub_district_en
          district_th: district_th
          province_en: province_en
          province_th: province_th
          sub_district_th: sub_district_th
          district_en: district_en
          zip_code: zip_code
        - sub_district_en: sub_district_en
          district_th: district_th
          province_en: province_en
          province_th: province_th
          sub_district_th: sub_district_th
          district_en: district_en
          zip_code: zip_code
        detail: detail
        title: title
        status: 0
    _internal_address_province_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/address.Province'
            type: array
        type: object
      example:
        data:
        - code: code
          province_en: province_en
          iso_code: iso_code
          province_th: province_th
        - code: code
          province_en: province_en
          iso_code: iso_code
          province_th: province_th
        detail: detail
        title: title
        status: 0
    _internal_bank_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/bank.Bank'
            type: array
        type: object
      example:
        data:
        - icon_url: icon_url
          code: code
          name: name
          name_th: name_th
          short_name: short_name
          id: id
        - icon_url: icon_url
          code: code
          name: name
          name_th: name_th
          short_name: short_name
          id: id
        detail: detail
        title: title
        status: 0
    _internal_bank_branch_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/bankBranch.BankBranch'
            type: array
        type: object
      example:
        data:
        - bank_code: bank_code
          bank_branch_code: bank_branch_code
          branch_name: branch_name
          id: id
        - bank_code: bank_code
          bank_branch_code: bank_branch_code
          branch_name: branch_name
          id: id
        detail: detail
        title: title
        status: 0
...